De-anonymize demultiplexing rules
The demultiplexing rules cannot have a name because they are defined in a
loop. This adds some code that modifies the Snakemake data structure in which
the rules are stored to set a sensible name afterwards.

This could break when the Snakemake data structures are changed, but since
this is just about aesthetics, we can just disable this code at that time.

Closes #5